+++ src/image-gallery/azext_image_gallery/vendored_sdks/azure_mgmt_compute/models/_models.py
@@ -1159,6 +1159,73 @@ class CommunityGallery(PirCommunityGalleryResource):
         super(CommunityGallery, self).__init__(**kwargs)
 
 
+class CommunityGalleryDiskImage(msrest.serialization.Model):
+    """This is the disk image base class.
+
+    Variables are only populated by the server, and will be ignored when sending a request.
+
+    :ivar size_in_gb: This property indicates the size of the VHD to be created.
+    :vartype size_in_gb: int
+    :param host_caching: The host caching of the disk. Valid values are 'None', 'ReadOnly', and
+     'ReadWrite'. Possible values include: "None", "ReadOnly", "ReadWrite".
+    :type host_caching: str or ~azure.mgmt.compute.v2021_07_01.models.HostCaching
+    """
+
+    _validation = {
+        'size_in_gb': {'readonly': True},
+    }
+
+    _attribute_map = {
+        'size_in_gb': {'key': 'sizeInGB', 'type': 'int'},
+        'host_caching': {'key': 'hostCaching', 'type': 'str'},
+    }
+
+    def __init__(
+        self,
+        **kwargs
+    ):
+        super(CommunityGalleryDiskImage, self).__init__(**kwargs)
+        self.size_in_gb = None
+        self.host_caching = kwargs.get('host_caching', None)
+
+
+class CommunityGalleryDataDiskImage(CommunityGalleryDiskImage):
+    """This is the data disk image.
+
+    Variables are only populated by the server, and will be ignored when sending a request.
+
+    All required parameters must be populated in order to send to Azure.
+
+    :ivar size_in_gb: This property indicates the size of the VHD to be created.
+    :vartype size_in_gb: int
+    :param host_caching: The host caching of the disk. Valid values are 'None', 'ReadOnly', and
+     'ReadWrite'. Possible values include: "None", "ReadOnly", "ReadWrite".
+    :type host_caching: str or ~azure.mgmt.compute.v2021_07_01.models.HostCaching
+    :param lun: Required. This property specifies the logical unit number of the data disk. This
+     value is used to identify data disks within the Virtual Machine and therefore must be unique
+     for each data disk attached to the Virtual Machine.
+    :type lun: int
+    """
+
+    _validation = {
+        'size_in_gb': {'readonly': True},
+        'lun': {'required': True},
+    }
+
+    _attribute_map = {
+        'size_in_gb': {'key': 'sizeInGB', 'type': 'int'},
+        'host_caching': {'key': 'hostCaching', 'type': 'str'},
+        'lun': {'key': 'lun', 'type': 'int'},
+    }
+
+    def __init__(
+        self,
+        **kwargs
+    ):
+        super(CommunityGalleryDataDiskImage, self).__init__(**kwargs)
+        self.lun = kwargs['lun']
+
+
 class CommunityGalleryImage(PirCommunityGalleryResource):
     """Specifies information about the gallery image definition that you want to create or update.
 
@@ -1288,6 +1355,12 @@ class CommunityGalleryImageVersion(PirCommunityGalleryResource):
     :param end_of_life_date: The end of life date of the gallery image version Definition. This
      property can be used for decommissioning purposes. This property is updatable.
     :type end_of_life_date: ~datetime.datetime
+    :param exclude_from_latest: If set to true, Virtual Machines deployed from the latest version
+     of the Image Definition won't use this Image Version.
+    :type exclude_from_latest: bool
+    :param storage_profile: This is the storage profile of a Gallery Image Version.
+    :type storage_profile:
+     ~azure.mgmt.compute.v2021_07_01.models.CommunityGalleryImageVersionStorageProfile
     """
 
     _validation = {
@@ -1303,6 +1376,8 @@ class CommunityGalleryImageVersion(PirCommunityGalleryResource):
         'unique_id': {'key': 'identifier.uniqueId', 'type': 'str'},
         'published_date': {'key': 'properties.publishedDate', 'type': 'iso-8601'},
         'end_of_life_date': {'key': 'properties.endOfLifeDate', 'type': 'iso-8601'},
+        'exclude_from_latest': {'key': 'properties.excludeFromLatest', 'type': 'bool'},
+        'storage_profile': {'key': 'properties.storageProfile', 'type': 'CommunityGalleryImageVersionStorageProfile'},
     }
 
     def __init__(
@@ -1312,6 +1387,8 @@ class CommunityGalleryImageVersion(PirCommunityGalleryResource):
         super(CommunityGalleryImageVersion, self).__init__(**kwargs)
         self.published_date = kwargs.get('published_date', None)
         self.end_of_life_date = kwargs.get('end_of_life_date', None)
+        self.exclude_from_latest = kwargs.get('exclude_from_latest', None)
+        self.storage_profile = kwargs.get('storage_profile', None)
 
 
 class CommunityGalleryImageVersionList(msrest.serialization.Model):
@@ -1342,6 +1419,30 @@ class CommunityGalleryImageVersionList(msrest.serialization.Model):
         super(CommunityGalleryImageVersionList, self).__init__(**kwargs)
         self.value = kwargs['value']
         self.next_link = kwargs.get('next_link', None)
+
+
+class CommunityGalleryImageVersionStorageProfile(msrest.serialization.Model):
+    """This is the storage profile of a Gallery Image Version.
+
+    :param os_disk_image: This is the OS disk image.
+    :type os_disk_image: ~azure.mgmt.compute.v2021_07_01.models.CommunityGalleryOSDiskImage
+    :param data_disk_images: A list of data disk images.
+    :type data_disk_images:
+     list[~azure.mgmt.compute.v2021_07_01.models.CommunityGalleryDataDiskImage]
+    """
+
+    _attribute_map = {
+        'os_disk_image': {'key': 'osDiskImage', 'type': 'CommunityGalleryOSDiskImage'},
+        'data_disk_images': {'key': 'dataDiskImages', 'type': '[CommunityGalleryDataDiskImage]'},
+    }
+
+    def __init__(
+        self,
+        **kwargs
+    ):
+        super(CommunityGalleryImageVersionStorageProfile, self).__init__(**kwargs)
+        self.os_disk_image = kwargs.get('os_disk_image', None)
+        self.data_disk_images = kwargs.get('data_disk_images', None)
 
 
 class CommunityGalleryInfo(msrest.serialization.Model):
@@ -1389,6 +1490,34 @@ class CommunityGalleryInfo(msrest.serialization.Model):
         self.public_name_prefix = kwargs.get('public_name_prefix', None)
         self.community_gallery_enabled = None
         self.public_names = None
+
+
+class CommunityGalleryOSDiskImage(CommunityGalleryDiskImage):
+    """This is the OS disk image.
+
+    Variables are only populated by the server, and will be ignored when sending a request.
+
+    :ivar size_in_gb: This property indicates the size of the VHD to be created.
+    :vartype size_in_gb: int
+    :param host_caching: The host caching of the disk. Valid values are 'None', 'ReadOnly', and
+     'ReadWrite'. Possible values include: "None", "ReadOnly", "ReadWrite".
+    :type host_caching: str or ~azure.mgmt.compute.v2021_07_01.models.HostCaching
+    """
+
+    _validation = {
+        'size_in_gb': {'readonly': True},
+    }
+
+    _attribute_map = {
+        'size_in_gb': {'key': 'sizeInGB', 'type': 'int'},
+        'host_caching': {'key': 'hostCaching', 'type': 'str'},
+    }
+
+    def __init__(
+        self,
+        **kwargs
+    ):
+        super(CommunityGalleryOSDiskImage, self).__init__(**kwargs)
 
 
 class ComputeOperationListResult(msrest.serialization.Model):
@@ -2139,9 +2268,11 @@ class DiagnosticsProfile(msrest.serialization.Model):
     """Specifies the boot diagnostic settings state. :code:`<br>`:code:`<br>`Minimum api-version: 2015-06-15.
 
     :param boot_diagnostics: Boot Diagnostics is a debugging feature which allows you to view
-     Console Output and Screenshot to diagnose VM status. :code:`<br>`:code:`<br>` You can easily
-     view the output of your console log. :code:`<br>`:code:`<br>` Azure also enables you to see a
-     screenshot of the VM from the hypervisor.
+     Console Output and Screenshot to diagnose VM status. :code:`<br>`\ **NOTE**\ : If storageUri is
+     being specified then ensure that the storage account is in the same region and subscription as
+     the VM. :code:`<br>`:code:`<br>` You can easily view the output of your console log.
+     :code:`<br>`:code:`<br>` Azure also enables you to see a screenshot of the VM from the
+     hypervisor.
     :type boot_diagnostics: ~azure.mgmt.compute.v2021_07_01.models.BootDiagnostics
     """
 
@@ -6160,10 +6291,8 @@ class RestorePoint(ProxyResource):
      https://aka.ms/RestorePoints for more details. Possible values include: "CrashConsistent",
      "FileSystemConsistent", "ApplicationConsistent".
     :vartype consistency_mode: str or ~azure.mgmt.compute.v2021_07_01.models.ConsistencyModeTypes
-    :ivar provisioning_details: Gets the provisioning details set by the server during Create
-     restore point operation.
-    :vartype provisioning_details:
-     ~azure.mgmt.compute.v2021_07_01.models.RestorePointProvisioningDetails
+    :param time_created: Gets the creation time of the restore point.
+    :type time_created: ~datetime.datetime
     """
 
     _validation = {
@@ -6173,7 +6302,6 @@ class RestorePoint(ProxyResource):
         'source_metadata': {'readonly': True},
         'provisioning_state': {'readonly': True},
         'consistency_mode': {'readonly': True},
-        'provisioning_details': {'readonly': True},
     }
 
     _attribute_map = {
@@ -6184,7 +6312,7 @@ class RestorePoint(ProxyResource):
         'source_metadata': {'key': 'properties.sourceMetadata', 'type': 'RestorePointSourceMetadata'},
         'provisioning_state': {'key': 'properties.provisioningState', 'type': 'str'},
         'consistency_mode': {'key': 'properties.consistencyMode', 'type': 'str'},
-        'provisioning_details': {'key': 'properties.provisioningDetails', 'type': 'RestorePointProvisioningDetails'},
+        'time_created': {'key': 'properties.timeCreated', 'type': 'iso-8601'},
     }
 
     def __init__(
@@ -6196,7 +6324,7 @@ class RestorePoint(ProxyResource):
         self.source_metadata = None
         self.provisioning_state = None
         self.consistency_mode = None
-        self.provisioning_details = None
+        self.time_created = kwargs.get('time_created', None)
 
 
 class RestorePointCollection(Resource):
@@ -6356,38 +6484,6 @@ class RestorePointCollectionUpdate(UpdateResource):
         self.provisioning_state = None
         self.restore_point_collection_id = None
         self.restore_points = None
-
-
-class RestorePointProvisioningDetails(msrest.serialization.Model):
-    """Restore Point Provisioning details.
-
-    :param creation_time: Gets the creation time of the restore point.
-    :type creation_time: ~datetime.datetime
-    :param total_used_size_in_bytes: Gets the total size of the data in all the disks which are
-     part of the restore point.
-    :type total_used_size_in_bytes: long
-    :param status_code: Gets the status of the Create restore point operation.
-    :type status_code: int
-    :param status_message: Gets the status message of the Create restore point operation.
-    :type status_message: str
-    """
-
-    _attribute_map = {
-        'creation_time': {'key': 'creationTime', 'type': 'iso-8601'},
-        'total_used_size_in_bytes': {'key': 'totalUsedSizeInBytes', 'type': 'long'},
-        'status_code': {'key': 'statusCode', 'type': 'int'},
-        'status_message': {'key': 'statusMessage', 'type': 'str'},
-    }
-
-    def __init__(
-        self,
-        **kwargs
-    ):
-        super(RestorePointProvisioningDetails, self).__init__(**kwargs)
-        self.creation_time = kwargs.get('creation_time', None)
-        self.total_used_size_in_bytes = kwargs.get('total_used_size_in_bytes', None)
-        self.status_code = kwargs.get('status_code', None)
-        self.status_message = kwargs.get('status_message', None)
 
 
 class RestorePointSourceMetadata(msrest.serialization.Model):
@@ -7498,7 +7594,7 @@ class SharingUpdate(msrest.serialization.Model):
     :param operation_type: Required. This property allows you to specify the operation type of
      gallery sharing update. :code:`<br>`:code:`<br>` Possible values are: :code:`<br>`:code:`<br>`
      **Add** :code:`<br>`:code:`<br>` **Remove** :code:`<br>`:code:`<br>` **Reset**. Possible values
-     include: "Add", "Remove", "Reset".
+     include: "Add", "Remove", "Reset", "EnableCommunity".
     :type operation_type: str or ~azure.mgmt.compute.v2021_07_01.models.SharingUpdateOperationTypes
     :param groups: A list of sharing profile groups.
     :type groups: list[~azure.mgmt.compute.v2021_07_01.models.SharingProfileGroup]