+++ src/aks-preview/HISTORY.md
@@ -4,7 +4,7 @@ Release History
 ===============
 0.5.41
 ++++++
-* Fix default value behavior for pod identity esception pod labels.
+* Fix default value behavior for pod identity exception pod labels.
 
 0.5.40
 +++++