+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,22 @@
 
 Release History
 ===============
+0.5.41
+++++++
+* Fix default value behavior for pod identity exception pod labels.
+
+0.5.40
++++++
+* Add support for new snapshot commands
+  * `az aks snapshot create`
+  * `az aks snapshot delete`
+  * `az aks snapshot list`
+  * `az aks snapshot show`
+* Add --snapshot-id to creating/upgrading commands
+  * `az aks create --snapshot-id`
+  * `az aks nodepool add --snapshot-id`
+  * `az aks nodepool upgrade --snapshot-id`
+
 0.5.39
 +++++
 * Add commands for agentpool start stop feature